Provide PreConfig and PostConfig lambdas to hook into sceptre
The "preconfig" lambda is invoked before the SC-Services stack and replaces the
"create_gophish_db.py" script.  This lambda creates the gophish database user
and database, and grants all access to the gophish database to the newly
created user.

The "postconfig" lambda is invoked after the SC-Services stack and replaces the
"update_api_parameter.py" script.  This lambda populates a secretsmanager entry
that contains the API key required to communicate with the gophish server.  The
status page relies on this API key to query campaign information.